Config mongoose, connect it, seed data to it
Added a configuration file which contains a function that establishes a connection to our MongoDB atlas database. In server.js we call this function and will henceforth retrieve our User and Product (and etc) information from a database. To give our database some starter data, we add seeding logic in seedData.js