Begin pages for Product/User listings/editing
These pages are all Admin protected. The Products and Users pages simply show a list of those resources (found from the Redux store), along with a button to delete / edit individual resources -- which is done on the Edit pages.

These are very ugly and very unstyled at the moment, and the code is very messy